requirements.txt
google-generativeai
PyMuPDF
genanki
python-dotenv
